Added new admin commands: * /adm rskip * /adm rres
and

* /cup res
* /cup skip

This will call dedicated methods RestartMap(false) and SkipMap(false);
so the cup scores will be cleared upon next map or restart.

I also added possibility to do the same as graphically.
The buttons for skip and restart can be found at force scores window.
There is also a button to reset scores without restart or skip.